requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
Farama-Notifications==0.0.4
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gymnasium==0.28.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax-jumpy==1.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
lightning-utilities==0.8.0
MarkupSafe==2.1.2
matplotlib==3.7.1
MinAtar==1.0.15
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.23
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
torch==2.0.1
torchmetrics==0.11.4
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
zipp==3.15.0
typer[all]==0.9.0
pydantic==2.0
